+++ pkNX.Structures/Dumpers/PersonalDumper.cs
@@ -65,7 +65,7 @@ namespace pkNX.Structures
         public IReadOnlyList<Learnset> EntryLearnsets { private get; set; }
         public IReadOnlyList<EggMoves> EntryEggMoves { private get; set; }
         public IReadOnlyList<EvolutionSet> Evos { private get; set; }
-        public IReadOnlyList<int> TMIndexes { protected get; set; }
+        public IReadOnlyList<ushort> TMIndexes { protected get; set; }
 
         private static readonly string[] AbilitySuffix = { " (1)", " (2)", " (H)" };
         private static readonly string[] ItemPrefix = { "Item 1 (50%)", "Item 2 (5%)", "Item 3 (1%)" };
